# Optimize VecDeque::append