@tabokie 2021-09-07
> Is Miri silent?

Sorry I'm new here, are you referring to running `./x.py test src/tools/miri`? If so the results look good.