@the8472 2021-09-10
I think that only runs the dedicated tests in the miri test directory. You can also run all `alloc` tests under miri. https://github.com/rust-lang/miri-test-libstd