@Mark-Simulacrum 2021-10-15
@bors r=m-ou-se

Yes, pushing code will invalidate the approved state.